fix: n1-1

+++ contracts/oracles/ChainlinkOracle.sol
@@ -140,7 +140,7 @@ contract ChainlinkOracle is AccessControlledV8, OracleInterface {
             price = _getChainlinkPrice(asset);
         }
 
-        uint256 decimalDelta = 18 - uint256(decimals);
+        uint256 decimalDelta = 18 - decimals;
         return price * (10 ** decimalDelta);
     }
